extract types into a common file
