+++ packages/blade/src/components/Table/componentIds.ts
@@ -1,4 +1,8 @@
 const ComponentIds = {
+  Table: 'Table',
+  TableBody: 'TableBody',
+  TableRow: 'TableRow',
+  TableCell: 'TableCell',
   TableToolbar: 'TableToolbar',
   TableHeader: 'TableHeader',
   TableHeaderRow: 'TableHeaderRow',
