+++ packages/blade/src/components/Table/index.ts
@@ -4,3 +4,4 @@ export * from './TableBody';
 export * from './TableFooter';
 export * from './TablePagination';
 export * from './TableToolbar';
+export * from './types';
